Summary - Patched ish to be able to see wire-frame drawing of zones   when compiled with 64 bit Sun Studio 12.  The critical   issues seem to be:   a) call to updwire() should be passed a local integer rather      than the common block value ncomp   b) calls to win3d() should have igw and igwh recalculated in      the fortran code after the call.   c) calls to axiscale should have the first two parameters passed      as local integers rather than the common block values.   d) in esru_x.c care is required when equating integers and      long integers. An explict cast is required.
  Several blocks of code in esrucom fitting this description and
  used by ish were patched.
  
  This commit to allow for testing on other machine types.
     
Testing
- Compile Sun Studio 12 on 64 bit machine generating 64 bit binary
  and then interactively use ish in graphics model as was able
  to view the wireframe after the patch.
- Compile GCC 3.4 X11 on 64 bit machine with -m32 directives.



Former-commit-id: cd005b8e975da899cdd2345a64449de5c1ae6219

+++ src/lib/esru_x.c
@@ -574,7 +574,7 @@ theVisual = XDefaultVisual(theDisp, theScreen);
 theCmap   = DefaultColormap(theDisp, theScreen);
 rootW     = RootWindow(theDisp,theScreen);
 dispDEEP  = DisplayPlanes(theDisp,theScreen);
-mdepth = dispDEEP;
+mdepth = (long int) dispDEEP;
 
 /* set colours for the border, background and forground */
 bd = WhitePixel(theDisp,theScreen);
@@ -961,7 +961,7 @@ char *act;  /* single character passed */
 long int *n;
 {
  int ic;
- ic = *n;
+ ic = (int) *n;
 /* sets the current forground colour n depending on which active colour set being used */
   if(*act == 'g') {
      if (ic >= 0 && ic <= ngscale ) {
@@ -1014,7 +1014,7 @@ long int *n;	/* index as in winscl_ */
 long int *xcolid;	/* index of colour used by X */
 {
  int ic;
- ic = *n;
+ ic = (int) *n;
 /* sets the current forground colour n depending on which active colour set being used */
   if(*act == 'g') {
      if (ic >= 0 && ic <= ngscale ) {
